Refactor scan and fix modules with registries
