+++ FixChain/modules/__init__.py
@@ -1,0 +1,1 @@
+"""Core modules package."""
