+++ FixChain/modules/fix/__init__.py
@@ -1,0 +1,6 @@
+from .registry import register, create
+
+# Ensure fixers are registered when package is imported
+from . import fixer  # noqa: F401
+
+__all__ = ["register", "create"]
