+++ FixChain/modules/rag/__init__.py
@@ -1,0 +1,3 @@
+from .registry import register, create
+
+__all__ = ["register", "create"]
