+++ FixChain/modules/rag/registry.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+from typing import Dict, Type
+
+from .base import RAG
+
+_registry: Dict[str, Type[RAG]] = {}
+
+
+def register(name: str, cls: Type[RAG]) -> None:
+    """Register a RAG class"""
+    _registry[name] = cls
+
+
+def create(name: str, *args, **kwargs) -> RAG:
+    """Create a RAG instance by name"""
+    if name not in _registry:
+        raise ValueError(f"RAG '{name}' is not registered")
+    return _registry[name](*args, **kwargs)
