+++ FixChain/modules/scan/__init__.py
@@ -1,0 +1,7 @@
+from .registry import register, create
+
+# Ensure scanners are registered when package is imported
+from . import bearer  # noqa: F401
+from . import sonar  # noqa: F401
+
+__all__ = ["register", "create"]
